+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-2025-10-08
-----------
+v1.1.0 (2025-10-09)
+--------------------
 
-- Version 1.0.0: core primitives, basic hints and support for environment variables
+- Added chained call methods: `Variable.given()` and `Variable.otherwise()`
+
+v1.0.0 (2025-10-08)
+-------------------
+
+- Core primitives, basic hints and support for environment variables